pyproject.toml
[project]
name = "web-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gunicorn>=23.0.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "req>=1.0.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "starlette>=0.50.0",
    "uuid7>=0.1.0",
    "uvicorn>=0.40.0",
]


[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["C", "E", "F", "W"]
ignore = ["C901", "F401", "F403", "E501"]
